[ITQT1-22] renamed the file processQuiz to calculateScore
