+++ apps/api/src/helper/calculateScore.ts
